% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MADFBA_functions.R
\name{plotExchMets}
\alias{plotExchMets}
\title{Plot active exchange reactions}
\usage{
plotExchMets(dfs, normalize = FALSE, main = "Concentrations")
}
\arguments{
\item{dfs}{optosol_dynamicFBA object from MADFBA simulation}

\item{normalize}{option to normalize data}
}
\description{
Plot active exchange reactions normalizing values if selected
}
